Simplified the implementation of Scenario listeners
ExampleTested now uses getScenario to return its ExampleNode (implementing
ScenarioInterface), like ScenarioTested, simplifying the implementation of
listener running for each scenario. They can depend on
AbstractScenarioTested without doing instanceof checks to get the
scenario.
